requirements.txt
langchain-core==0.1.0
langchain-community==0.1.0
langchain-text-splitters==0.1.0
langchain-openai==0.1.0
python-dotenv==1.0.0
faiss-cpu==1.7.2
